apps/backend/requirements.txt
flask
flask-cors
pydantic
requests
python-dotenv
Flask-SQLAlchemy>=3.1.1
openai